+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## Unreleased
+## UNRELEASED
 
 * Adds translation strings.
+* Use `self.apos.root.import` instead of `self.apos.root.require`.
+* `enablePassportStrategies` is now async.
 
 ## 1.2.0 - 2023-06-08
 
@@ -48,4 +50,3 @@ Declared stable. No code changes.
 ## 1.0.0-beta - 2022-01-06
 
 Initial release for A3. Tested and working with Google and Okta. Other standard passport modules should also work, especially those based on OpenAuth.
-